# apps/compass-ai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Compass AI Backend API",
    "version": "2.1.0",
    "description": "Document analysis with grounded, citation-backed answers, scoped to a client portal. Tenancy is derived from a verified portalFeatureContextToken sent as x-portal-context: NO operation accepts a client, workspace or portal identifier from the caller, and one that carries one is rejected with TENANCY_PROBE. Every operation enforces the derived actor (employee or client) inside its handler. All operations are private: there is no cross-app consumer."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Every failure carries a stable machine code. For a portal failure the message is separately authored to be safe to show a client.",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "examples": [
                  "TENANCY_PROBE",
                  "PORTAL_NOT_BOUND",
                  "FORBIDDEN_CLIENT_ROLE",
                  "PORTAL_VERIFY_FAILED"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "queued",
          "uploading",
          "ocr",
          "parsed",
          "embedding",
          "indexed",
          "failed",
          "deleted"
        ]
      },
      "DocumentSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "statusDetail": {
            "type": [
              "string",
              "null"
            ]
          },
          "errorMessage": {
            "type": [
              "string",
              "null"
            ]
          },
          "pageCount": {
            "type": [
              "integer",
              "null"
            ]
          },
          "chunkCount": {
            "type": "integer"
          },
          "sourceKind": {
            "type": "string",
            "enum": [
              "native",
              "ocr",
              "hybrid"
            ]
          },
          "needsOcr": {
            "type": "boolean"
          },
          "ocrPagesUsed": {
            "type": "integer"
          },
          "byteSize": {
            "type": "integer"
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "indexedAt": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "chunkCount",
          "byteSize",
          "version",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
    },
    "parameters": {
      "PortalContext": {
        "name": "x-portal-context",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Platform-signed portalFeatureContextToken from the iframe URL. Verified through Gate on every request; a local decode is never trusted. It proves WHERE the request is from, never WHO — identity comes from the caller session."
      },
      "DocumentId": {
        "name": "documentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "TenancyProbe": {
        "description": "The request carried a client/workspace/portal identifier. Refused rather than ignored.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PortalRejected": {
        "description": "Portal context missing, unverifiable, unbound, paused, or the caller is not a member.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ForbiddenClient": {
        "description": "FORBIDDEN_CLIENT_ROLE — this action is employee-only.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found, or it belongs to another client. Deliberately indistinguishable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/documents": {
      "get": {
        "operationId": "listDocuments",
        "summary": "List the documents of this portal",
        "description": "Scoped by the verified portal. Clients see only indexed documents; employees also see processing and failed ones, because they are the people who fix them.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          }
        ],
        "responses": {
          "200": {
            "description": "Documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DocumentSummary"
                      }
                    },
                    "ocrConfigured": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "documents"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      },
      "post": {
        "operationId": "uploadDocuments",
        "summary": "Upload PDFs",
        "description": "Employee-only. The client id comes from the portal context and never from the form; a tenant key in the form is a TENANCY_PROBE. Ingestion runs on the queue, so this returns 202.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "async",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          }
        ],
        "responses": {
          "202": {
            "description": "Per-file outcome"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "files"
                ]
              }
            }
          }
        }
      }
    },
    "/documents/{documentId}": {
      "get": {
        "operationId": "getDocument",
        "summary": "Read one document",
        "description": "A document of another client is a 404, never a 403.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Document"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      },
      "delete": {
        "operationId": "deleteDocument",
        "summary": "Delete a document and its index entries",
        "description": "Employee-only and irreversible: pages, paragraphs and passages cascade from the document row.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      }
    },
...
  }
}